requirements.txt
# Core Trading Dependencies
kiteconnect>=4.2.0
pandas>=2.1.4
numpy>=1.26.2

# Technical Analysis
pandas-ta>=0.4.67b0
ta>=0.11.0

# Data Management
gspread>=5.12.4
oauth2client>=4.1.3

# API and Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-multipart>=0.0.6

# Notifications
python-telegram-bot>=20.7
requests>=2.31.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging and Monitoring
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
schedule>=1.2.0

# Data Visualization (for backtesting reports)
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0

# Database (optional, for trade history)
sqlalchemy>=2.0.25

# Performance (optional - skip if installation fails)
# numba>=0.58.1
